% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_a}
\alias{cohort_a}
\title{Cohort A}
\format{
A data frame with 10,000 rows and 9 columns
}
\source{
Simulated data
}
\usage{
cohort_a
}
\description{
Example dataset of cohort data
}
\keyword{datasets}
