% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_b}
\alias{cohort_b}
\title{Cohort B}
\format{
A data frame with 5,000 rows and 5 columns
}
\source{
Simulated data
}
\usage{
cohort_b
}
\description{
Example dataset of cohort data
}
\keyword{datasets}
