% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_c}
\alias{cohort_c}
\title{Cohort C}
\format{
A data frame with 7,000 rows and 5 columns
}
\source{
Simulated data
}
\usage{
cohort_c
}
\description{
Example dataset of cohort data
}
\keyword{datasets}
