% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_shell_function.R
\name{cohort_shell_func}
\alias{cohort_shell_func}
\title{Cohort sheet create. To be called by harmonization function.}
\usage{
cohort_shell_func(sheet)
}
\arguments{
\item{sheet}{Harmonization sheet}
}
\value{

}
\description{
Cohort sheet create. To be called by harmonization function.
}
\examples{

}
