% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_or_cat_function.R
\name{cont_or_cat}
\alias{cont_or_cat}
\title{Continuous or categorical}
\usage{
cont_or_cat(data, var)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable}
}
\value{
Returns "continuous" or "categorical"
}
\description{
Continuous or categorical
}
\examples{

# Function can help determine which kind of output is
# most appropriate

cont_or_cat(data = cohort_a, var = 'height_1')

cont_or_cat(data = cohort_a, var = 'education')

}
