% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_log_report.R
\name{create_error_log_report}
\alias{create_error_log_report}
\title{Error log report creation}
\usage{
create_error_log_report(
  harmonization_object,
  path = "./",
  file = "error_log_report.html"
)
}
\arguments{
\item{harmonization_object}{Harmonization object}

\item{path}{Path of output R Markdown report}

\item{file}{Filename of output R Markdown report}
}
\description{
This function will create an RMarkdown error log. It takes the harmonization
object as the input, and will knit an RMarkdown html file to the path
specified.

\strong{Note:} The error log will only be able to detect "processing" errors, and
not "content" errors. For example, if the user enters coding instructions
that are nonsensical or incorrect, but are still able to be executed, this
function will not be able to detect it.
}
