% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Summary report creation}
\usage{
create_summary_report(
  harmonization_object,
  path = "./",
  file = "summary_report.html",
  compare = FALSE
)
}
\arguments{
\item{harmonization_object}{Harmonization object}

\item{path}{Path of output R Markdown report}

\item{file}{Filename of output R Markdown report}

\item{compare}{Creates summary report with comparison of raw values with modified values}
}
\value{
Does not return an object, but instead knits html RMarkdown report to specified path and file name.
}
\description{
Summary report creation
}
\examples{

# Examples not run

# Creating harmonized object
# harmonized_obj <- harmonization(harmonization_sheet = harmonization_sheet_example)

# Knitting summary report
# create_summary_report(harmonization_object = harmonized_obj,
#   path = './',
#   file = 'example_output.html)

# Use `compare` option to create comparison summary report.
# create_summary_report(harmonization_object = harmonized_obj,
#   path = './',
#   file = 'example_output.html,
#   compare = TRUE)


}
