% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_harmonization_sheet_example}
\alias{error_harmonization_sheet_example}
\title{Error harmonization sheet example}
\format{
A data frame with 16 rows and 12 columns
}
\source{
Created data
}
\usage{
error_harmonization_sheet_example
}
\description{
Example of harmonization sheet. This harmonization sheet has a few typos
present (incorrect name of variable and/or dataset). This can be used to
demonstrate how errors are presented in the error log and/or the summary
method.
}
\keyword{datasets}
