% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonization_sheet_example}
\alias{harmonization_sheet_example}
\title{Harmonization sheet example}
\format{
A data frame with 16 rows and 12 columns
}
\source{
Created data
}
\usage{
harmonization_sheet_example
}
\description{
Example of a harmonization sheet. This serves as the input file for the harmonization function.
}
\keyword{datasets}
