% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_function.R
\name{range_function}
\alias{range_function}
\title{Range function. To be called by harmonization function.}
\usage{
range_function(
  data = temp_dataset,
  min_max_range = possible_range,
  new_var = item
)
}
\arguments{
\item{data}{Data to be modified}

\item{min_max_range}{Range of allowed values}

\item{new_var}{New variable}
}
\value{
Returns a list with the new vector (values outside of range set to NA), and the number of values set to NA.
}
\description{
Range function. To be called by harmonization function.
}
\examples{

test_data <- data.frame(val = 1:10)

range_function(data = test_data, min_max_range = '[2,8]', new_var = 'val')

}
