% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_function.R
\name{range_function_cat}
\alias{range_function_cat}
\title{Possible values for categorical variables. To be called by harmonization function.}
\usage{
range_function_cat(
  data = temp_dataset,
  possible_vals_cat = possible_vals,
  new_var = item
)
}
\arguments{
\item{data}{data to be modified}

\item{possible_vals_cat}{vector of possible values}

\item{new_var}{new variable}
}
\description{
Possible values for categorical variables. To be called by harmonization function.
}
