% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_reorder_function.R
\name{reorder_factors_df}
\alias{reorder_factors_df}
\title{Reorder factors data.frame}
\usage{
reorder_factors_df(data, sheet)
}
\arguments{
\item{data}{Harmonized data.frame}

\item{sheet}{Factor reorder sheet}
}
\value{
Returns harmonized data.frame.
}
\description{
Reorder factors data.frame
}
\examples{

}
