% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psHarmonize_summary_method.R
\name{summary.psHarmonize}
\alias{summary.psHarmonize}
\title{psHarmonize summary method}
\usage{
\method{summary}{psHarmonize}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{psHarmonize object}

\item{...}{Can provide additional arguments}

\item{verbose}{T/F. When TRUE, will list variables for each section.}
}
\value{
Doesn't return object. Prints status of harmonization (# of harmonizated variables, etc.)
}
\description{
psHarmonize summary method
}
\examples{

harmonization_obj <- harmonization(harmonization_sheet_example)

summary(harmonization_obj)

# Use verbose option to see more details
summary(harmonization_obj, verbose = TRUE)

}
