% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_modify_func.R
\name{code_modify_func}
\alias{code_modify_func}
\title{Code modify function. To be called by the harmonization function.}
\usage{
code_modify_func(
  data = temp_dataset,
  instruction = code_instruct,
  old_var = source_item_long
)
}
\arguments{
\item{data}{Dataframe to be modified}

\item{instruction}{Coding instruction from harmonization sheet}

\item{old_var}{Name of original variable}
}
\value{

}
\description{
Code modify function. To be called by the harmonization function.
}
\examples{

}
