% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_modify_func_multi.R
\name{code_modify_func_multi}
\alias{code_modify_func_multi}
\title{Code modify function (multiple variables). To be called by the harmonization function.}
\usage{
code_modify_func_multi(
  data = temp_dataset,
  instruction = code_instruct,
  old_var = source_item_long,
  user_args = source_item,
  sourcedataset = source_dataset,
  subdomain = subdomain,
  visit = visit,
  cohort = cohort
)
}
\arguments{
\item{data}{Dataframe to be modified}

\item{instruction}{Coding instruction from harmonization sheet}

\item{old_var}{Name of original variable}

\item{user_args}{Character vector of input variables}

\item{sourcedataset}{Dataframe created so far}

\item{subdomain}{Category of variable}

\item{visit}{Visit number}

\item{cohort}{Cohort name}
}
\description{
Code modify function (multiple variables). To be called by the harmonization function.
}
