% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_shell_function.R
\name{cohort_shell_func}
\alias{cohort_shell_func}
\title{Cohort sheet create. To be called by harmonization function.}
\usage{
cohort_shell_func(sheet)
}
\arguments{
\item{sheet}{Harmonization sheet}
}
\value{
Data.frame with IDs, study/cohort name, and visit.
}
\description{
Created dataframe "shell" of IDs, study/cohort name, and visit. Harmonized
variables will be joined onto this dataset.
}
\examples{

# Using example harmonization sheet
cohort_shell_func(harmonization_sheet_example)

}
