% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_log_report.R
\name{create_error_log_report}
\alias{create_error_log_report}
\title{Error log report creation}
\usage{
create_error_log_report(
  harmonization_object,
  path = "./",
  file = "error_log_report.html"
)
}
\arguments{
\item{harmonization_object}{Harmonization object}

\item{path}{Path of output R Markdown report}

\item{file}{Filename of output R Markdown report}
}
\value{

}
\description{
Error log report creation
}
\examples{

}
