% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_dataset.R
\name{create_long_dataset}
\alias{create_long_dataset}
\title{Create long dataset. To be called by harmonization function.}
\usage{
create_long_dataset(
  vars_interest,
  subdomain,
  previous_dataset,
  error_log,
  na_string
)
}
\arguments{
\item{vars_interest}{Variable currently being harmonized}

\item{subdomain}{Category of variable}

\item{previous_dataset}{Dataframe created so far}

\item{error_log}{Error log}

\item{na_string}{Character string of final recode value to be set to NA.}
}
\value{

}
\description{
Create long dataset. To be called by harmonization function.
}
\examples{

}
