% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Summary report creation}
\usage{
create_summary_report(
  harmonization_object,
  path = "./",
  file = "summary_report.html",
  compare = FALSE
)
}
\arguments{
\item{harmonization_object}{Harmonization object}

\item{path}{Path of output R Markdown report}

\item{file}{Filename of output R Markdown report}

\item{compare}{Creates summary report with comparison of raw values with modified values}
}
\value{

}
\description{
Summary report creation
}
\examples{

}
