% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_function.R
\name{range_function_cat}
\alias{range_function_cat}
\title{Possible values for categorical variables. To be called by harmonization function.}
\usage{
range_function_cat(
  data = temp_dataset,
  possible_vals_cat = possible_vals,
  new_var = item
)
}
\arguments{
\item{data}{data to be modified}

\item{possible_vals_cat}{vector of possible values}

\item{new_var}{new variable}
}
\value{
Returns a list with the new vector (values outside of set to NA), and the number of values set to NA.
}
\description{
Possible values for categorical variables. To be called by harmonization function.
}
\examples{

test_data <- data.frame(val = c('a','b','j','k','c','d'))

range_function_cat(data = test_data, possible_vals_cat = c('a','b','c','d'), new_var = 'val')

}
