% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psHarmonize_summary_method.R
\name{summary.psHarmonize}
\alias{summary.psHarmonize}
\title{psHarmonize summary method}
\usage{
\method{summary}{psHarmonize}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{psHarmonize object}

\item{...}{Can provide additional arguments}

\item{verbose}{T/F. When TRUE, will list variables for each section.}
}
\description{
psHarmonize summary method
}
