% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psHarmonize_summary_method.R
\name{summary.psHarmonize}
\alias{summary.psHarmonize}
\title{psHarmonize summary method}
\usage{
\method{summary}{psHarmonize}(object, verbose = F)
}
\arguments{
\item{verbose}{}
}
\description{
psHarmonize summary method
}
